+++ documents/changes.md
@@ -3,6 +3,7 @@
 This is a list of changes between the old binding api and the new
 
 1. crsbinding -> crs.binding
+2. crsbinding._exfn -> crs.binding.functions
 
 ## Extensions